requirements.txt
click==7.1.2
Flask==1.1.2
Flask-HTTPAuth==4.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pytube3==9.6.4
typing-extensions==3.7.4.2
Werkzeug==1.0.1
